Bugfix: support variable keylength
The constructor-set `keyLength` instance variable was being used during encryption (`digest` method) but not during matching `decodeAndCheckMatches` function); instead, in the latter, a hardcoded `keyLength` value of `32` was always being used to generate the SCrypt-derived key.  This would cause permanent failure for SCrypt-encoded password hashes that were generated by an SCryptPasswordEncoder initialized with a `keyLength` != 32.

This change abstracts the call of the BouncyCastle `SCrypt.digest` method in order to prevent the DRY-principal-violation-facilitated error that caused the discrepancy by always passing in the constructor-intitialized instance variable config values to the `SCrypt.digest` method.  By doing so, support for keyLengths!= 32 is introduced.